% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rklib.r
\name{rk.groupby}
\alias{rk.groupby}
\title{Process GroupBy Clause}
\usage{
rk.groupby(groupby = NULL, mode = "@")
}
\arguments{
\item{groupby}{clause following simplified sql syntax.}

\item{mode}{specifies the parsing logic. The default value '@' returns the groupby clause in perfmeter format. Any other value returns the groupby fields used for aggregation.}
}
\value{
The function returns the processed groupby clause.
}
\description{
The function parses the groupby clause and returns the modified string as per the specified mode.
}
