% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rklib.r
\name{rk.metadata}
\alias{rk.metadata}
\title{Processing Meta Data}
\usage{
rk.metadata(table = NULL, fields = NULL, gettabs, getflds,
  infofile = NULL)
}
\arguments{
\item{table}{name of the interface tables. If not specified, all the tables of the given interface are returned.}

\item{fields}{is the list of field names whose details are required. If not specified, all the fields of the specified tables are returned.}

\item{gettabs}{is a function that returns the list of all the tables for the given interface.}

\item{getflds}{is a function that returns the list of all the fields for the given table.}

\item{infofile}{is the name of the file containing information about different tables and fields of the given interface.}
}
\value{
The function returns the resulting data frame.
}
\description{
The function returns the list of tables, fields, and their descriptions.
}
