% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.r
\name{rk.query}
\alias{rk.query}
\title{Data Processing Query}
\usage{
rk.query(dframe, fields = NULL, where = NULL, groupby = NULL)
}
\arguments{
\item{fields}{to be filtered.}

\item{where}{clause applied on the data.}

\item{groupby}{used to aggregate the fields.}

\item{data}{frame to be processed.}
}
\value{
The function returns the resulting data frame.
}
\description{
The function applies the given fields, where clause, and group by fields on the specified data frame.
}
\examples{
data <- rk.query(data, "CpuPctAvg,MemPct", "PlatformId = 0 AND ComponentId = 11")
data <- rk.query(data, "ComponentId,AVG(CpuPctAvg),MAX(MemPct)", "PlatformId = 0", "ComponentId")
}
\seealso{
\code{\link{rk.metadata}} for tables and field details, \code{\link{rk.fields}} for field clause, \code{\link{rk.where}} for where clause and \code{\link{rk.groupby}} for groupby clause.
}
