% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njira.r
\name{jira.login}
\alias{jira.login}
\title{Jira Login Function}
\usage{
jira.login(jira.env = NULL, jira.user = NULL, jira.pwd = NULL,
  jira.val = 0, logs = FALSE)
}
\arguments{
\item{jira.env}{Web address of 'Jira' environment (e.g. https://issues.apache.org/jira)}

\item{jira.user}{Jira User Name}

\item{jira.pwd}{Jira Password}

\item{jira.val}{0/1 how should the list values be returned in the query results}

\item{logs}{debug logs required on not (Default = FALSE)}
}
\value{
The function authenticates the user into Jira installation and caches the Jira credentials.
}
\description{
Authenticates the user to fetch data from the respective Jira installation.
}
\examples{
jira.login(jira.env="https://issues.apache.org/jira", 
jira.user="jiraTestUser", jira.pwd="jiraTestPwd")

}
