% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/njira.r
\name{jira.login}
\alias{jira.login}
\title{Jira Login Function}
\usage{
jira.login(
  jira.env = NULL,
  jira.user = NULL,
  jira.pwd = NULL,
  jira.val = 0,
  logs = F
)
}
\arguments{
\item{jira.env}{Web address of JIRA environment (e.g. https://issues.apache.org/jira)}

\item{jira.user}{Jira User Name}

\item{jira.pwd}{Jira Password}

\item{jira.val}{0/1 how should the list values be returned in the query results}

\item{logs}{set it to True or False based on loggin is required on not (Default = F)}
}
\value{
The function autheticates into JIRA environment..
}
\description{
This function authenticates the user to fetch data from respective JIRA environment.
}
\examples{
jira.login(jira.env="https://issues.apache.org/jira", 
jira.user="jiraTestUser", jira.pwd="jiraTestPwd")
}
