% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rklib.r
\name{rk.query}
\alias{rk.query}
\title{Data Processing Query}
\usage{
rk.query(dframe, fields = NULL, where = NULL, groupby = NULL)
}
\arguments{
\item{dframe}{data frame to be processed.}

\item{fields}{fields to be filtered.}

\item{where}{clause applied on the data.}

\item{groupby}{used to aggregate the fields.}
}
\value{
The function returns the resulting data frame.
}
\description{
The function applies the given fields, where clause, and group by fields on the specified data frame.
}
